requirements.txt
opencv-python~=4.9.0.80
numpy~=1.24.3
matplotlib~=3.7.5
scipy~=1.10.1
pillow~=10.2.0
mmpose~=1.3.1
pandas~=2.0.3
torchvision~=0.17.2
scikit-learn~=1.3.0